+++ src/scene-tree.cpp
@@ -79,7 +79,7 @@ void SceneTree::resizeEvent(QResizeEvent *event)
 		}
 
 		int wid = contentsRect().width() - scrollWid - 1;
-		int items = (int)std::ceil((float)wid / maxWidth);
+		int items = (int)std::ceil((float)wid / (float)maxWidth);
 		int itemWidth = wid / items;
 
 		setGridSize(QSize(itemWidth, itemHeight));
@@ -123,14 +123,14 @@ void SceneTree::dropEvent(QDropEvent *event)
 			setVerticalScrollBarPolicy(Qt::ScrollBarAlwaysOn);
 		}
 
-		float wid = contentsRect().width() - scrollWid - 1;
+		float wid = float(contentsRect().width() - scrollWid - 1);
 
 		QPoint point = event->position().toPoint();
 
-		int x = (float)point.x() / wid * std::ceil(wid / maxWidth);
+		int x = int((float)point.x() / wid * std::ceil(wid / (float)maxWidth));
 		int y = (point.y() + firstItemY) / itemHeight;
 
-		int r = x + y * std::ceil(wid / maxWidth);
+		int r = x + y * (int)std::ceil(wid / (float)maxWidth);
 
 		QListWidgetItem *item = takeItem(selectedIndexes()[0].row());
 		insertItem(r, item);
@@ -163,15 +163,15 @@ void SceneTree::RepositionGrid(QDragMoveEvent *event)
 		setVerticalScrollBarPolicy(Qt::ScrollBarAlwaysOn);
 	}
 
-	float wid = contentsRect().width() - scrollWid - 1;
+	float wid = float(contentsRect().width() - scrollWid - 1);
 
 	if (event) {
 		QPoint point = event->position().toPoint();
 
-		int x = (float)point.x() / wid * std::ceil(wid / maxWidth);
+		int x = int((float)point.x() / wid * std::ceil(wid / (float)maxWidth));
 		int y = (point.y() + firstItemY) / itemHeight;
 
-		int r = x + y * std::ceil(wid / maxWidth);
+		int r = x + y * (int)std::ceil(wid / (float)maxWidth);
 		int orig = selectedIndexes()[0].row();
 
 		for (int i = 0; i < count(); i++) {
@@ -186,8 +186,8 @@ void SceneTree::RepositionGrid(QDragMoveEvent *event)
 				  (i > orig && i > r ? 1 : 0) -
 				  (i > orig && i == r ? 2 : 0);
 
-			int xPos = (i + off) % (int)std::ceil(wid / maxWidth);
-			int yPos = (i + off) / (int)std::ceil(wid / maxWidth);
+			int xPos = (i + off) % (int)std::ceil(wid / (float)maxWidth);
+			int yPos = (i + off) / (int)std::ceil(wid / (float)maxWidth);
 			QSize g = gridSize();
 
 			QPoint position(xPos * g.width(), yPos * g.height());
@@ -202,8 +202,8 @@ void SceneTree::RepositionGrid(QDragMoveEvent *event)
 
 			QModelIndex index = indexFromItem(wItem);
 
-			int xPos = i % (int)std::ceil(wid / maxWidth);
-			int yPos = i / (int)std::ceil(wid / maxWidth);
+			int xPos = i % (int)std::ceil(wid / (float)maxWidth);
+			int yPos = i / (int)std::ceil(wid / (float)maxWidth);
 			QSize g = gridSize();
 
 			QPoint position(xPos * g.width(), yPos * g.height());
